✨ feat(): created new Title component, organized codes
