+++ src/components/Input.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { PropTypes } from "prop-types";
-import { FormsInput, FormsLabel, InputGroup } from "../styles/Input";
+import { FormsInput, FormsLabel, InputGroup } from "../styles/components/Input";
 
 export default function Input(props) {
   const { inputType, inputName, inputAutocomplete, inputOnChange, textLabel } =
