+++ src/components/Title.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { PropTypes } from "prop-types";
+import { PrimaryTitle } from "../styles/components/Title";
+
+export default function Title(props) {
+  const { titleText } = props;
+  return <PrimaryTitle>{titleText}</PrimaryTitle>;
+}
+
+Title.propTypes = {
+  titleText: PropTypes.string,
+}.isRequired;
