+++ src/styles/components/FormsRegister.js
@@ -11,12 +11,6 @@ export const Content = styled.section`
   width: 360px;
 `;
 
-export const Text = styled.span`
-  color: ${(props) => (props.textColor === "red" ? "#e32636" : "#f0f8ff")};
-  margin: 1rem;
-  text-align: center;
-`;
-
 export const Link = styled.a`
   margin: 1rem;
 `;
