+++ src/styles/components/Logo.js
