+++ src/styles/components/Title.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const PrimaryTitle = styled.h1`
+  color: #f0f8ff;
+  margin: 1rem;
+  text-align: center;
+  font-size: 2rem;
+  font-weight: 700;
+  letter-spacing: 0.125rem;
+  line-height: 3rem;
+`;
