+++ src/styles/page/Register.js
